# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everwin AI Platform - Enterprise Control Plane",
    "description": "Enterprise Multi-tenant AI Gateway & Inference Middleware Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/nlp/translation": {
      "post": {
        "tags": [
          "Natural Language Processing (Translation & Summarization)"
        ],
        "summary": "Bilingual Text Translation (Helsinki-NLP)",
        "operationId": "nlp_translation_v1_nlp_translation_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nlp/summarization": {
      "post": {
        "tags": [
          "Natural Language Processing (Translation & Summarization)"
        ],
        "summary": "Intelligent Text Summarization",
        "operationId": "nlp_summarization_v1_nlp_summarization_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SummarizationRequest": {
        "properties": {
          "document": {
            "type": "string",
            "title": "Document",
            "example": "Văn bản tài liệu dài cần tóm tắt..."
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio",
            "description": "Tỷ lệ tóm tắt văn bản",
            "default": 0.2
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "SummarizationRequest"
      },
...
      "TranslationRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "example": "Xin chào thế giới, chào mừng đến với Everwin AI Platform."
          },
          "source_lang": {
            "type": "string",
            "title": "Source Lang",
            "default": "vi",
            "example": "vi"
          },
          "target_lang": {
            "type": "string",
            "title": "Target Lang",
            "default": "en",
            "example": "en"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TranslationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}